Mass renaming.
Rename a bunch of files to more reader-friendly filenames,
since they're going to appear in the book. Rename "program_structure_"
files to "crates_". Other random changes.

These are now in sync with the book: for every fully reproduced error
message in the book, we have an antisample, and you can get the error by
just `cd`-ing into the `compile-fail` directory and typing:

    rustc whatever_file.rs